add extension loader for fetchers
